move select_values to the json_ext module
